lib/ukrandom: Rename ukswrand to ukrandom
This commit renames ukswrand library to ukrandom to better describe its
use case. This library will provide cryptographically secure random
numbers, either directly from the hardware, or from a PRNG seeded with
the true random seed.

swrand.h is the main header of the library and should match its name,
therefore it is renamed to random.h. swrand.c is also renamed to match
the header.

Signed-off-by: Alex Apostolescu <alexx.apostolescu@gmail.com>