lib/posix-poll: Fix finalizer duplication in epoll
Commit 7b2e38171 (lib/posix-poll: Autoremove closed files from epoll)
introduced a logic error that would register a file finalizer on every
call to EPOLL_CTL_MOD. This lead to use-after-free errors when a file
was removed from epoll.
This change is a quick fix to this error. A more elegant refactoring of
epoll code should be considered when vfscore shim is no longer required.

Signed-off-by: Andrei Tatar <andrei@unikraft.io>