lib/ukrandom: Clean the public API
This commit addresses the public API by renaming the main randomness
function uk_swrand_fill_buffer() to uk_random_fill_buffer(), to abide
Unikraft's naming convention.

As a result, the private hwrng function fill_buffer() is renamed to
uk_random_hwrng_fill_buffer().

uk_random_fill_buffer() is the result of a PRNG and it never fails.
Make it void to reflect this.

Signed-off-by: Alex Apostolescu <alexx.apostolescu@gmail.com>