lib/ukrandom: Add option to seed the CSPRNG with a fixed value
Add an option to seed the CSPRNG with a fixed value. This options is
intended for development purposes, or running Unikraft in legacy
systems that don't provide a TRNG.

Enabling this option compromises the system's security. The build
system will mark Unikraft as tainted and libukrandom will print a
banner during init.

DO NOT ENABLE THIS OPTIION IN PRODUCTION SYSTEMS.

Signed-off-by: Michalis Pappas <michalis@unikraft.io>