lib/ukboot: Added heap randomization
When LIBUKBOOT_HEAP_BASE and RUNTIME_ASLR are set, a random
virtual address for the heap is created, which overrides the
value from LIBUKBOOT_HEAPBASE. The addrerss is generated between
the last virtual base from the last memory region and 2 ** 46.
The upper limit is set to 2 ** 46 due to the canonical addresses
used in x86-64 architecture.

When LIBUKBOOT_HEAP_BASE is disabled, the logic is similar with
the stack randomization. A random base address is generated inside
any of the remaining memory regions. The rest of them are added
to the allocator.

Signed-off-by: Serban Sorohan <serban.sorohan@gmail.com>