support/scripts: Added scripts used for compile-time randomization
We have three new scripts used for the compile-time randomization.
'base_address.py' is used to randomize the base address of the
'.text' section.
'Analyzer.py' is used to parse the linker file 'link64.lds' and
extract the sections which will be randomized
'ASLR.py' is used to randomize the sections extracted by 'Analyzer.py',
randomize the order of PHDRS, randomize the order of libraries, and
add padding between any two adjacent libs and write the new
'link64_ASLR.lds' file.

Signed-off-by: Serban Sorohan <serban.sorohan@gmail.com>