plat/kvm: Move `serial_console.c` into `drivers/ukconsole/ns16550/`
`serial_console.c` implements a driver for the COM ports
on x86. The driver is now moved out of the platform and
will be registered with `ukconsole` in the future. The device
underlying the COM port interface is a 16550 / 8250 so the
`ns16550` subdirectory is the right place to put this driver.
Because the device is the same on Arm and on x86, the `com.c`
and `ns16550.c` driver should be merged into one driver in
the future.

Signed-off-by: Thassilo Schulze <thassilo@unikraft.io>