arch: Add exception context escape trampolines ctx creation
Implement the ability to jump from exception handling context to an
a context where one can take exceptions or yield the current thread and
from which one can return to the context that the
exception handler was supposed to return to.

By calling `ukarch_ctx_init_ehtrampo`, caller passes a usable stack
pointer as one of the arguments where the general purpose registers
along with what else makes the execution envirionments will be stored
as well as the argument passed by the caller. The caller passed
callback will be guaranteed to be in a flexible context where it can do
everything it wants given the guarantee that upon exit, it will return
safely to the context that generated the exception in the first place.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>