lib/ukconsole: Add abstraction layer for console drivers
`ukconsole` is a new library that implements a generic
interface for console devices. It will be used to replace
the current kernel/debug print system that's based on the
`ukplat_c[out|in][k|d]` functions. `ukconsole` is
groundwork for properly integrating the gdb stub with
Unikraft and for adding a proper kernel print system
later on.

Signed-off-by: Thassilo Schulze <thassilo@unikraft.io>