lib/ukfalloc: Export allocator stats
Export system-wide and per-allocator statistics via libukstore. This
is controlled by a hidden Kconfig option selected by frame allocator
libraries that wish to export metrics.

Entry IDs of the exported metrics are provided in include/uk/falloc_store.h

Since the frame allocator is initialized before the heap, early calls
to uk_falloc_init_stats() are deferred to init. This is not visible to
the caller.

Signed-off-by: Michalis Pappas <michalis@unikraft.io>