drivers/uktty/ns16550: Move MMIO functions into own module
The NS16550 on ARM is the same device that underlies the
COM ports on x86. That means we can use the NS16550 driver
on both architectures. One difference between ARM and x86
is that we use MMIO to access the device on ARM and port
IO to access the device on x86. Another difference is that
ARM uses the device-tree to get information about the device
whereas x86 just uses well-known, fixed locations.

This commit moves the MMIO-specific part of the NS16550 driver
into a separate module. This module implements read/write
operations for NS16550 device registers. The interface
implemented by the MMIO module can later be implemented for
x86 using port IO. This way, we can share common code.

Checkpatch-Ignore: MACRO_ARG_REUSE
Signed-off-by: Thassilo Schulze <thassilo@unikraft.io>