lib/ukdebug: Introduce gdb stub beginning with send/receive functions
This commit starts introducing the gdb stub with the functions that
allow the stub to send data to and receive data from the gdb host.
The actual transmission of data isn't implemented yet. That's why
the `gdb_dbg_putc` and `gdb_dbg_getc` functions are stubs for now.

Checkpatch-Ignore: MACRO_WITH_FLOW_CONTROL
Signed-off-by: Marc Rittinghaus <marc.rittinghaus@unikraft.io>
Reviewed-by: Michalis Pappas <michalis@unikraft.io>
Reviewed-by: Simon Kuenzer <simon@unikraft.io>
Approved-by: Simon Kuenzer <simon@unikraft.io>
GitHub-Closes: #1479