lib/ukdebug: Add memory read and write packets to gdb stub
These packets allow the gdb host to read and write memory on the
target system. This capability is used to implement breakpoints for
example.

There's an edge-case that the implementation of `gdb_arch_write_memory`
must handle. Assume virtual memory is enabled with `ukvmem` and consider
the following sequence of events:

1. the system starts and the gdb stub is entered early on in the boot
   sequence
2. the gdb stub writes a breakpoint to some location in the program's
   text
3. execution is resumed. Now virtual memory is initialized. The location
   in the program's text that the stub wrote to is marked read-only
4. the breakpoint that the stub created is hit. The stub has to write
   to the read-only page to remove the breakpoint and resume exeuction

This edge-case is handled by using addresses from `ukplat_page_kmap`
instead of the addresses that are passed to `gdb_arch_write_memory`.

Checkpatch-Ignore: UNNECESSARY_PARENTHESES
Co-authored-by: Thassilo Schulze <thassilo@unikraft.io>
Signed-off-by: Thassilo Schulze <thassilo@unikraft.io>
Signed-off-by: Marc Rittinghaus <marc.rittinghaus@unikraft.io>