drivers/uktty: Move NS16550 FDT code into MMIO module
The NS16550 driver has a feature matrix that looks something
like this:

+--------+------+---------+
|        | MMIO | port IO |
+--------+------+---------+
| FDT    | ARM  |         |
+--------+------+---------+
| No FDT |      | x86     |
+--------+------+---------+

Idealy, we'd have an FDT module, a "no FDT" module an MMIO
module and a port IO module. Then any one architecture could
be implemented by picking the appropriate modules.

But in reality, the features aren't as indendent as this feature
matrix implies. E.g., the existing FDT code specifically deals with
MMIO-related values. Similarly, the "no FDT" code from the x86 COM
driver is x86-specific. So the features in the matrix cannot be
decoupled completely.

This commit moves the FDT helper functions from the main NS16550
driver into the MMIO module. The helper functions are also updated
slightly.

Signed-off-by: Thassilo Schulze <thassilo@unikraft.io>