drivers/xenheaders: Move public Xen headers to own library
This commit moves the official public Xen API headers to a separate
library. These headers are originally copy from the Xen source tree
and define the API to the hypervisor and virtual devices. This is done
to better organize Xen-related code and to have single place for
updates and dependency resolutions.
The created library is kept invisible because it is only needed by Xen
drivers and it is itself not a driver (reduce confusion).

Signed-off-by: Simon Kuenzer <simon@unikraft.io>