lib/ukrandom: Add initialization via the cmdline
Add a new Kconfing option, namely LIBUKRANDOM_CMDLINE_INIT
to seed the CSPRNG via the cmdline. In ChaCha this is used
to pass 8 bytes for the key followed by 2 bytes for the nonce
in `random.seed` cmdline parameter.

When this option is enabled AND initialziation parameters are
passed to the cmdline, the underlying RNG driver is not used.

lib/ukrandom: Seed the CSPRNG from the cmdline

Provide the ability to seed ChaCha20 from the kernel's command
line. This is controlled by a newly introduced Kconfig option,
CONFIG_LIBUKRANDOM_CMDLINE_SEED. The relevant cmdline parameter
is `random.seed` that should by populted by the loader with an
array of 8 random 32-bit unsigned integers for the ChaCha20 key.

Seeding via the cmdline takes precedence over the device-tree and
drivers to RNG devices.

Checkpatch-Ignore: SPACING
Signed-off-by: Michalis Pappas <michalis@unikraft.io>