drivers/ns16650: Rename `libukconsole_ns16550` to `libns16550`
Rename the driver library and driver directory to `libns16550` to provide
a naming scheme that is more appropriate for general purpose devices. The
name `ns16550` is considered a precise description of the driver as it is
the name of the actual serial I/O chip.
Configuration options are adopted accordingly.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>