drivers/ukconsole/ns16550: Fix Kconfig warnings on invalid `choice`
Fix Kconfig warnings on ns16550 for x86_64 caused by choice groups
defined without an associated symbol:

> warning: choice value used outside its choice group

According to [1]:

"
choices:
	"choice" [symbol]
	<choice options>
	<choice block>
	"endchoice"
...

If no [symbol] is associated with a choice, then you can not have multiple
definitions of that choice. If a [symbol] is associated to the choice,
then you may define the same choice (ie. with the same entries) in another
place."

[1] https://android.googlesource.com/kernel/common/+/bcmdhd-3.10/Documentation/kbuild/kconfig-language.txt

Signed-off-by: Michalis Pappas <michalis@unikraft.io>