lib/posix-fdtab: Allow named files & partial open
This change expands the fdtab API, separating the action of creating a
new open file description with that of associating it with an fd. This
allows callers to perform additional initialization on the open file
description before the fd goes live.

One notable such init is filling in the `name` field of the ofile,
which the API additions now support and take care to allocate space for.

It was a conscious decision to not mandate that the fdtab fill in the
name itself, as drivers may construct names in any manner of ways other
than having a string on hand. Thus, to prevent a redundant memcpy, a
driver can choose to fill in the field itself.

Signed-off-by: Andrei Tatar <andrei@unikraft.io>