plat/xen: Include fp_arm64.c in FPSIMD Xen ARM64 builds
`fp_arm64.c` implements `fpsimd_save_state()` and
`fpsimd_restore_state()` that are required by FPSIMD ARM64 builds. The
file was not included in the build rules (in `Makefile.uk`) resulting in
a build error.

Fix the build error by including `fp_arm64.c` in the Xen build rules.
This is similar to its inclusion in the KVM build rules.

GitHub-Fixes: #1522
Signed-off-by: Razvan Deaconescu <razvand@unikraft.io>
Reviewed-by: Oleksii Moisieiev <oleksii_moisieiev@epam.com>
Approved-by: Michalis Pappas <michalis@unikraft.io>
GitHub-Closes: #1523