lib/vfscore: Support embedded initrd with custom compiled-in fstabs
This commit enables compiling images with embedded initrds with "custom"
compiled-in mount tables (fstabs). As soon as one of the activated
compiled-in mount table references to the device `"embedded"`, the
configuration requires a path for an initrd for embedding at compile
time. Prio this change, an initrd could only be embedded via using
pre-configured (and non-"custom") table configuration that depends on
an embedded initrd, for example: "Embedded InitRD (CPIO)".

Signed-off-by: Simon Kuenzer <simon@unikraft.io>