lib/syscall-shim: Do not dereference optional clone() parameters
The parent_tid parameter of clone() is used to instruct the kernel
where to store the child TID in parent's memory. Similarly, the
child_tid is used to instruct the kernel where to store the child
TID in the child's memory. Both parameters are optional, and are
interpreted conditionally to whether the CLONE_CHILD_SETTID and
CLONE_PARENT_SETTID flags are set, respectively.

Do not interpret these options as PT_REF, as the pointer will not
be valid if the caller does not set corresponding flags.

Checkpatch-Ignore: LONG_LINE
Signed-off-by: Michalis Pappas <michalis@unikraft.io>