lib/posix-process: Add per-process `sbrk`/`brk` implementations
Implement support for per-process `sbrk`/`brk`. This feature has two
variants: one that uses libukvmem and one that alternatively simply
allocates a big enough area from the global allocator.

In both cases, since we have to track per process the brk base and
its current position this makes the feature depend on there being an
associated `struct posix_process` with the thread upon initialization.

Additionally, one is able to choose the size of the backing memory area
by configuring `CONFIG_LIBPOSIX_PROCESS_BRK_PAGE_ORDER` (default 2MiB).

Another important note is that, by default, we adopt the behavior of
glibc with MORECORE_CLEARS == 2, meaning that whenever there is a bump
through (s)brk, we zero out that difference.
This increment zeroing behavior is configurable through
`LIBPOSIX_PROCESS_BRK_INC_MEMZERO`.

For the non-ukvmem variant we simply request a big enough area from the
global allocator and just move the brk cursor around based on requests,
while taking care to zero out on bumps.

For the ukvmem variant, we initialize an anonymous, demand mapped, VMA.
Unlike the previous variant, this one does not interfere with the global
allocator.

As far as maintaining `MORECORE_CLEARS == 2` behavior goes, every newly
faulted page will be guaranteed to be zeroed out by the page fault
handler due to the VMA not having the UK_VMA_MAP_UNINITIALIZED flag.
On one hand, this helps us in not having to memset big, page boundary
crossing, bumps all at once and, instead, have the page fault handler
zero them out as they are being used. On the other hand, we must still
manually take care of in-page bumps: if the bump does not cross a page
boundary, it will not fault and thus not be zeroed out, forcing us to do
it ourselves.

What is more, the ukvmem variant is able to free up pages upon brk
decrement requests, if the difference is a multiple of page size.
This behavior is configurable through
`LIBPOSIX_PROCESS_BRK_DEC_DONTNEED`.
This is POSIX-compliant as POSIX specifies that the pages upon
reincrement have undefined values but remains optional as it is not
Linux-compatible since there the pages do not get released and thus
reincremented pages would keep previous data (although some libcs can
do manual zeroing).

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>