lib/posix-process/signals: Fix conditional for sigsetops
The implementation of sigsetops was conditional to HAVE_LIBC instead of
CONFIG_HAVE_LIBC, causing a build error when musl is used, due to the
definiton of sigset_t to different types. Update the conditional to fix.
Notice that this works as each flavor of libc controls both the type
and the implementation of the functions that operate on it.

Signed-off-by: Michalis Pappas <michalis@unikraft.io>