lib/posix-process: Add uk_posix_process_run()
This function spanws a process that jumps to main-like function passed
as an argument. It is intended to be used by the init process when
multiprocess is enabled, to spawn the application process and call
main(). As in native mode there is no path to pass to execve(), this
function allows spawning the application process in both native and
bincompat mode.

Signed-off-by: Michalis Pappas <michalis@unikraft.io>