lib/vfscore: Rework signature of vfscore_lseek
This change reworks vfscore_lseek to be more in line with its syscall as
well as the other vfscore I/O functions, using a unified return value as
opposed to dedicated error and offset outputs.
Calling code has been patched to reflect this change.
Vfscore diff kept to a minimum, as it's soon to be deprecated.

Signed-off-by: Andrei Tatar <andrei@unikraft.io>