lib/posix-fdio: Add sendfile syscall
This change adds support for the sendfile syscall, supporting any
combination of legacy and ukfiles.
In case of ukfile-ukfile transfer, dedicated sendfile is used.
In all other cases, a fallback implementation using two I/O calls and a
temporary stack buffer is used, shimming between vfscore and posix-fdio
as necessary.

Signed-off-by: Andrei Tatar <andrei@unikraft.io>