lib/posix-process: Terminate underlying thread only if added to sched
The pthread termination path may try to terminate the underlying thread
if it's different from the current one. However this implies that it
has been added to the scheduler at some point and this may not be true
if, for example, it is simply being released on the error/cleanup path
of some function that created said thread and hasn't gotten to adding
it to a scheduler yet.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>