lib/syscall_shim/arch/arm64: Properly restore `x30` in prologue
Commit 87fa095ae8bf ("lib/syscall_shim/arch/arm64: Store caller's link register")
updated the way we save the link register to properly emulate a
system call trap, so that the actual link register substitued our
supposed ELR_EL1 and the link register from before substitutes the
actual link register of the function that trapped. In order to achieve
the latter it does a level 1 stack unwinding to extract the link
register saved at offset 8 from the frame pointer.

However, the commit forgot to also update the way we restore the link
register as now it wrongly restores the link register saved from the
unwinding instead of the one from where the syscall was called.

Fix this by properly restoring the link register from where we
actually save it now after said commit.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>